+++ packages/core-js/modules/es.array.iterator.js
@@ -55,7 +55,7 @@ addToUnscopables('keys');
 addToUnscopables('values');
 addToUnscopables('entries');
 
-// V8 ~ Chrome 43-45 bug
+// V8 ~ Chrome 45- bug
 if (!IS_PURE && DESCRIPTORS && values.name !== 'values') try {
   defineProperty(values, 'name', { value: 'values' });
 } catch (error) { /* empty */ }